Some #include improvements
