+++ include/supernode/requests/sale_details.h
@@ -1,7 +1,7 @@
 
 #pragma once
 
-#include "lib/graft/inout.h"
+#include "lib/graft/serialize.h"
 #include "lib/graft/router.h"
 
 #include <vector>
